Rename package and merge related files
This CR addresses the remaining comments in the CR
(https://gerrit.hyperledger.org/r/#/c/21775/)

Primarily - renames the package valinternal to internal
and merges the two files with related content into one.

FAB-12189 #done
