+++ core/ledger/kvledger/txmgmt/validator/internal/tx_ops_preparation.go
@@ -4,7 +4,7 @@ Copyright IBM Corp. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package valinternal
+package internal
 
 import (
 	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/privacyenabledstate"
@@ -152,4 +152,76 @@ func retrieveLatestMetadata(ns, coll, key string,
 		return vv.Metadata, nil
 	}
 	return db.GetPrivateDataMetadataByHash(ns, coll, []byte(key))
+}
+
+type keyOpsFlag uint8
+
+const (
+	upsertVal      keyOpsFlag = 1 // 1 << 0
+	metadataUpdate            = 2 // 1 << 1
+	metadataDelete            = 4 // 1 << 2
+	keyDelete                 = 8 // 1 << 3
+)
+
+type compositeKey struct {
+	ns, coll, key string
+}
+
+type txOps map[compositeKey]*keyOps
+
+type keyOps struct {
+	flag     keyOpsFlag
+	value    []byte
+	metadata []byte
+}
+
+////////////////// txOps functions
+
+func (txops txOps) upsert(k compositeKey, val []byte) {
+	keyops := txops.getOrCreateKeyEntry(k)
+	keyops.flag += upsertVal
+	keyops.value = val
+}
+
+func (txops txOps) delete(k compositeKey) {
+	keyops := txops.getOrCreateKeyEntry(k)
+	keyops.flag += keyDelete
+}
+
+func (txops txOps) metadataUpdate(k compositeKey, metadata []byte) {
+	keyops := txops.getOrCreateKeyEntry(k)
+	keyops.flag += metadataUpdate
+	keyops.metadata = metadata
+}
+
+func (txops txOps) metadataDelete(k compositeKey) {
+	keyops := txops.getOrCreateKeyEntry(k)
+	keyops.flag += metadataDelete
+}
+
+func (txops txOps) getOrCreateKeyEntry(k compositeKey) *keyOps {
+	keyops, ok := txops[k]
+	if !ok {
+		keyops = &keyOps{}
+		txops[k] = keyops
+	}
+	return keyops
+}
+
+////////////////// keyOps functions
+
+func (keyops keyOps) isDelete() bool {
+	return keyops.flag&(keyDelete) == keyDelete
+}
+
+func (keyops keyOps) isUpsertAndMetadataUpdate() bool {
+	if keyops.flag&upsertVal == upsertVal {
+		return keyops.flag&metadataUpdate == metadataUpdate ||
+			keyops.flag&metadataDelete == metadataDelete
+	}
+	return false
+}
+
+func (keyops keyOps) isOnlyUpsert() bool {
+	return keyops.flag|upsertVal == upsertVal
 }
