+++ core/ledger/kvledger/txmgmt/validator/valimpl/default_impl.go
@@ -12,8 +12,8 @@ import (
 	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/privacyenabledstate"
 	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/txmgr"
 	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/validator"
+	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/validator/internal"
 	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/validator/statebasedval"
-	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/validator/valinternal"
 )
 
 var logger = flogging.MustGetLogger("valimpl")
@@ -21,11 +21,11 @@ var logger = flogging.MustGetLogger("valimpl")
 // DefaultImpl implements the interface validator.Validator
 // This performs the common tasks that are independent of a particular scheme of validation
 // and for actual validation of the public rwset, it encloses an internal validator (that implements interface
-// valinternal.InternalValidator) such as statebased validator
+// internal.InternalValidator) such as statebased validator
 type DefaultImpl struct {
-	txmgr txmgr.TxMgr
-	db    privacyenabledstate.DB
-	valinternal.InternalValidator
+	txmgr             txmgr.TxMgr
+	db                privacyenabledstate.DB
+	internalValidator internal.Validator
 }
 
 // NewStatebasedValidator constructs a validator that internally manages statebased validator and in addition
@@ -39,8 +39,8 @@ func (impl *DefaultImpl) ValidateAndPrepareBatch(blockAndPvtdata *ledger.BlockAn
 	doMVCCValidation bool) (*privacyenabledstate.UpdateBatch, error) {
 	block := blockAndPvtdata.Block
 	logger.Debugf("ValidateAndPrepareBatch() for block number = [%d]", block.Header.Number)
-	var internalBlock *valinternal.Block
-	var pubAndHashUpdates *valinternal.PubAndHashUpdates
+	var internalBlock *internal.Block
+	var pubAndHashUpdates *internal.PubAndHashUpdates
 	var pvtUpdates *privacyenabledstate.PvtUpdateBatch
 	var err error
 
@@ -49,7 +49,7 @@ func (impl *DefaultImpl) ValidateAndPrepareBatch(blockAndPvtdata *ledger.BlockAn
 		return nil, err
 	}
 
-	if pubAndHashUpdates, err = impl.InternalValidator.ValidateAndPrepareBatch(internalBlock, doMVCCValidation); err != nil {
+	if pubAndHashUpdates, err = impl.internalValidator.ValidateAndPrepareBatch(internalBlock, doMVCCValidation); err != nil {
 		return nil, err
 	}
 	logger.Debug("validating rwset...")
